change get_target_path functionality
