+++ lang/docs/src/lib.rs
@@ -4,5 +4,4 @@ pub mod generate_docs;
 pub mod util;
 
 pub use doc::write_html;
-pub use util::get_target_path;
 pub use util::open;
